kinda fixed some of the dom/ local storage but then it isn't fully working
