+++ src/scripts/submitButton.js
@@ -1,5 +1,8 @@
 let taskListFunction = require("./taskList")
 let databaseFunction = require("./database");
+let dom = require("./domBuilder");
+const allTasks = databaseFunction.pullFromStorage("ArrayOfTasks");
+
 function inputForm () {
     const todo= document.getElementById("todoInput").value;
     const dueDay= document.getElementById("daydropdown").value
@@ -8,17 +11,10 @@ function inputForm () {
     const dueDate = dueDay + " " + dueMonth + ", " + dueYear;
     const description = document.getElementById("todoDescription").value;
     taskListFunction.createTask(todo, description, dueDate, "to do");
-    // const header = document.createTextNode(todo)
-    // const text = document.createTextNode(description);
-    // const newItem = document.createElement("li")
-    // console.log(newItem, "New Item")
-    // newItem.appendChild(header);
-    // newItem.appendChild(text);
-    // document.getElementById("todoList").appendChild(newItem)
-
   }
-document.getElementById("submit-button").addEventListener("click", inputForm)
-
-
-taskListFunction.createTask("Say hello", "I'm dexcribing here", "04/15/2018", "to do");
+// document.getElementById("submit-button").addEventListener("click", inputForm)
+document.getElementById("submit-button").addEventListener("click", () => {
+  inputForm();
+  dom();
+});
