+++ src/scripts/taskList.js
@@ -37,9 +37,9 @@ const taskList = Object.create({}, {
         }
     }
 })
-taskList.createTask("get happy", "coding stressing me out", "2/2/2018", "to do");
-taskList.createTask("new pet", "get pet salamander", "2/2/2019", "to do");
-console.log(taskList.listArray)
+// taskList.createTask("get happy", "coding stressing me out", "2/2/2018", "to do");
+// taskList.createTask("new pet", "get pet salamander", "2/2/2019", "to do");
+// console.log(taskList.listArray)
 
 //console.log("list array", listArray)
 //databaseFunction.saveDatabase(taskList.listArray, "ArrayOfTasks")
